rename minDifi file
Fixes 908
