+++ src/sorcha/modules/PPMiniDifi.py
